+++ tools/perf/builtin-c2c.h
@@ -5,6 +5,7 @@
 #include <linux/list.h>
 #include <stdbool.h>
 #include <stdint.h>
+#include <inttypes.h>
 
 #include "ui/browsers/hists.h"
 #include "util/mem-events.h"
@@ -16,6 +17,69 @@
 #include "util/env.h"
 #include "util/map.h"
 #include "util/maps.h"
+#include "util/sort.h"
+
+/* Constants */
+#define SYMBOL_WIDTH 30
+#define C2C_HEADER_MAX 2
+
+/**
+ * struct c2c_header - Column header definition for C2C display
+ * @line: Array of header lines with text and span
+ */
+struct c2c_header {
+	struct {
+		const char *text;
+		int	    span;
+	} line[C2C_HEADER_MAX];
+};
+
+/**
+ * struct c2c_dimension - Definition of a display column for C2C
+ * @header: Column header text and span
+ * @name: Column name for configuration
+ * @width: Default column width
+ * @se: Sort entry if this dimension uses standard sorting
+ * @cmp: Comparison function for sorting
+ * @entry: Entry rendering function
+ * @color: Colored entry rendering function (optional)
+ */
+struct c2c_dimension {
+	struct c2c_header	 header;
+	const char		*name;
+	int			 width;
+	struct sort_entry	*se;
+
+	int64_t (*cmp)(struct perf_hpp_fmt *fmt,
+		       struct hist_entry *, struct hist_entry *);
+	int   (*entry)(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		       struct hist_entry *he);
+	int   (*color)(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		       struct hist_entry *he);
+};
+
+/**
+ * struct c2c_fmt - Format wrapper for C2C dimensions
+ * @fmt: Base perf HPP format structure
+ * @dim: Pointer to the C2C dimension
+ */
+struct c2c_fmt {
+	struct perf_hpp_fmt	 fmt;
+	struct c2c_dimension	*dim;
+};
+
+/* Helper macros for string formatting */
+#define HEX_STR(__s, __v)				\
+({							\
+	scnprintf(__s, sizeof(__s), "0x%" PRIx64, __v);	\
+	__s;						\
+})
+
+#define PERC_STR(__s, __v)				\
+({							\
+	scnprintf(__s, sizeof(__s), "%.2F%%", __v);	\
+	__s;						\
+})
 
 /**
  * struct compute_stats - Statistics computed from memory access samples
@@ -214,6 +278,82 @@ struct perf_c2c {
 
 /* Global C2C context - defined in builtin-c2c.c */
 extern struct perf_c2c c2c;
+
+/* Dimension declarations - defined in builtin-c2c.c */
+extern struct c2c_dimension dim_symbol;
+extern struct c2c_dimension dim_srcline;
+
+/**
+ * symbol_width - Calculate width for symbol column
+ * @hists: Histogram context
+ * @se: Sort entry for symbol
+ *
+ * Returns: Column width respecting SYMBOL_WIDTH limit
+ */
+int symbol_width(struct hists *hists, struct sort_entry *se);
+
+/**
+ * c2c_width - Calculate width for a C2C column
+ * @fmt: HPP format
+ * @hpp: HPP context
+ * @hists: Histogram context
+ *
+ * Returns: Column width based on dimension configuration
+ */
+int c2c_width(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp, struct hists *hists);
+
+/**
+ * he_stats - Get C2C statistics for a histogram entry
+ * @he: Histogram entry
+ *
+ * Returns: Pointer to C2C stats for this entry
+ */
+struct c2c_stats *he_stats(struct hist_entry *he);
+
+/**
+ * total_stats - Get total C2C statistics for histogram
+ * @he: Histogram entry
+ *
+ * Returns: Pointer to total C2C stats for the histogram
+ */
+struct c2c_stats *total_stats(struct hist_entry *he);
+
+/**
+ * percent - Calculate percentage
+ * @st: Part value
+ * @tot: Total value
+ *
+ * Returns: Percentage as double
+ */
+static inline double percent(u32 st, u32 tot)
+{
+	return tot ? 100. * (double) st / (double) tot : 0;
+}
+
+/* Macro for calculating percentage of a field */
+#define PERCENT(__h, __f) percent(he_stats(__h)->__f, total_stats(__h)->__f)
+
+/* Entry functions for symbol view - defined in c2c-symbol.c */
+int total_stores_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		       struct hist_entry *he);
+int cacheline_symbol_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+			   struct hist_entry *he);
+int symbol_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		 struct hist_entry *he);
+int cycles_rmt_hitm_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+			  struct hist_entry *he);
+int cycles_lcl_hitm_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+			  struct hist_entry *he);
+int cycles_load_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		      struct hist_entry *he);
+int cycles_total_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+		       struct hist_entry *he);
+int cnt_other_load_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+			 struct hist_entry *he);
+int cycles_percent_entry(struct perf_hpp_fmt *fmt, struct perf_hpp *hpp,
+			 struct hist_entry *he);
+int64_t cycles_percent_cmp(struct perf_hpp_fmt *fmt,
+			   struct hist_entry *left, struct hist_entry *right);
 
 /**
  * struct c2c_symbol_browser - Symbol browser for C2C analysis
